// src/lib/api/models/createFieldDefinitionDtoFieldType.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * RealHub API
 * Multi-tenant Real Estate Platform API
 * OpenAPI spec version: v1
 */

export type CreateFieldDefinitionDtoFieldType = typeof CreateFieldDefinitionDtoFieldType[keyof typeof CreateFieldDefinitionDtoFieldType];


export const CreateFieldDefinitionDtoFieldType = {
  TEXT: 'TEXT',
  TEXTAREA: 'TEXTAREA',
  NUMBER: 'NUMBER',
  MONEY: 'MONEY',
  BOOLEAN: 'BOOLEAN',
  DATE: 'DATE',
  DATETIME: 'DATETIME',
  SELECT: 'SELECT',
  MULTI_SELECT: 'MULTI_SELECT',
  RADIO: 'RADIO',
  CHECKBOX: 'CHECKBOX',
  FILE: 'FILE',
  IMAGE: 'IMAGE',
  LOCATION: 'LOCATION',
  JSON: 'JSON',
} as const;